update rule template controller. (#74)

+++ octopus-api/src/main/java/org/metahut/octopus/api/dto/MetricsConfigResponseDTO.java
@@ -44,10 +44,10 @@ public class MetricsConfigResponseDTO {
     private Date updateTime;
 
     @ApiModelProperty(value = "creator")
-    private Integer creator;
+    private Long creator;
 
     @ApiModelProperty(value = "updater")
-    private Integer updater;
+    private Long updater;
 
     public Integer getId() {
         return id;
@@ -137,19 +137,19 @@ public class MetricsConfigResponseDTO {
         this.updateTime = updateTime;
     }
 
-    public Integer getCreator() {
+    public Long getCreator() {
         return creator;
     }
 
-    public void setCreator(Integer creator) {
+    public void setCreator(Long creator) {
         this.creator = creator;
     }
 
-    public Integer getUpdater() {
+    public Long getUpdater() {
         return updater;
     }
 
-    public void setUpdater(Integer updater) {
+    public void setUpdater(Long updater) {
         this.updater = updater;
     }
 }
